tests: Moved SwitchFlowTableTest to switch_test
flow_table_test.SwitchFlowTableTest is now tested under the name of
switch_test.ProcessFlowModTest. This is due to the moving of the method
process_flow_mod from flow_table to switch, making it previously
incompatible with the unit tests.

Since the unit tests are under the switches instead, the tests are
modified to do the same as the other tests, using a mock controller to
send across flow mod messages rather than directly process them as done
beforehand.