datapaths: Refactor/rewrite flow mod handling
This is an attempt to split the previously clutter _process_flow_mod
into multiple pieces, in order to reduce duplicated code and improve
flexibility for future possible flow mod messages. Some method names
were changed to keep it separate from that of others with the same
prefix originally. The parameter strict is now made an argument passed
to modification and deletion flow mod handling to reduce checking of
the command type.